requirements.txt
absl-py==2.4.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
filelock==3.28.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
idna==3.11
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.33
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==4.25.9
psutil==7.2.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
